# jobportal-docs/apis/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "description": "Retrieve the profile of the authenticated user",
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get Current User Profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Update the authenticated user's profile information",
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              },
              "examples": {
                "updateEmail": {
                  "summary": "Update Email",
                  "value": {
                    "email": "newemail@example.com"
                  }
                },
                "updatePassword": {
                  "summary": "Update Password",
                  "value": {
                    "password": "NewSecurePass123!"
                  }
                },
                "updateProfile": {
                  "summary": "Update Full Profile",
                  "value": {
                    "name": "John Updated",
                    "email": "john.updated@example.com",
                    "address": "789 New Street, Boston, MA 02101",
                    "telephone": "+1-555-999-8888"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update User Profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me/upload-resume": {
      "post": {
        "description": "Upload a resume file to the user profile (max 5MB)",
        "operationId": "UsersController_uploadResume",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Resume uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resume uploaded successfully",
                    "resumeUrl": "https://cloudinary.com/..."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file format or size"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upload Resume",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{userId}": {
      "get": {
        "description": "Retrieve a user profile by their ID",
        "operationId": "UsersController_getUserProfile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID (MongoDB ObjectId)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get User Profile by ID",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Job Portal API",
    "description": "A comprehensive API for managing jobs, applications, users, and messaging. Supports multiple user roles: JobSeeker, Employer, and Admin.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "JobSeeker",
              "Employer",
              "Admin"
            ]
          },
          "resumeUrl": {
            "type": "string",
            "example": "https://cloudinary.com/resume.pdf"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "address": {
            "type": "string",
            "example": "123 Main St, New York, NY"
          },
          "telephone": {
            "type": "string",
            "example": "+1-555-0199"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "email",
          "role",
          "name",
          "address",
          "telephone",
          "isActive"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "newemail@example.com"
          },
          "password": {
            "type": "string",
            "example": "newpassword123",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "JobSeeker",
              "Employer",
              "Admin"
            ]
          }
        }
      },
...
    }
  }
}